Introduce SysInfoBase class and fix childProcess spawnFromEvent
Due to differencies in the way AppBuilder CLI and NativeScript CLI work with android tools move the SysInfo class to SysInfoBase.
This is required in case when `--log trace` is used, for example:
`tns doctor --log trace` - in this case, the dispatcher (part of common lib) is collecting system information immediately when the process starts.
The information is cached and when DoctorService tries to get the information with correct android and adb paths, they are not used as result is already cached.
This leads to different results when executing `tns doctor` and `tns doctor --log trace`.
So introduce SysInfoBase class and allow the CLIs to extend it with their own implementation.

Also fix spawnFromEvent method in childProcess as it should not throw error when the throwError parameter is false.
